api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-jose[cryptography]==3.3.0
